+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const listContacts = require('./listContacts')
+const getContactById = require('./getContactById')
+const addContact = require('./addContact')
+const removeContact = require('./removeContact')
+const updateContact = require('./updateContact')
+
+module.exports = {
+  listContacts,
+  getContactById,
+  addContact,
+  removeContact,
+  updateContact,
+}
