+++ controllers/contacts/listContacts.js
@@ -1,5 +1,5 @@
-const contactsOperations = require('../model')
-const { sendSuccessRes } = require('../helpers')
+const contactsOperations = require('../../model/contacts')
+const { sendSuccessRes } = require('../../helpers')
 
 const listContacts = async (req, res, next) => {
   const result = await contactsOperations.listContacts()
