+++ db/contacts.json
@@ -1,0 +1,62 @@
+[
+  {
+    "id": "1c344192-b88d-4696-ba0b-d99260866a09",
+    "name": "Allen Raymond",
+    "email": "nulla.ante@vestibul.co.uk",
+    "phone": "(992) 914-3792"
+  },
+  {
+    "id": "1c344192-b88d-4696-ba0b-d99260867a01",
+    "name": "Chaim Lewis",
+    "email": "dui.in@egetlacus.ca",
+    "phone": "(294) 840-6685"
+  },
+  {
+    "id": 3,
+    "name": "Kennedy Lane",
+    "email": "mattis.Cras@nonenimMauris.net",
+    "phone": "(542) 451-7038"
+  },
+  {
+    "id": "1c344192-b88d-4696-ba0b-d99260867a02",
+    "name": "Wylie Pope",
+    "email": "est@utquamvel.net",
+    "phone": "(692) 802-2949"
+  },
+  {
+    "id": "1c344192-b88d-4696-ba0b-d99260867a03",
+    "name": "Cyrus Jackson",
+    "email": "nibh@semsempererat.com",
+    "phone": "(501) 472-5218"
+  },
+  {
+    "id": "1c344192-b88d-4696-ba0b-d99260867a04",
+    "name": "Abbot Franks",
+    "email": "scelerisque@magnis.org",
+    "phone": "(186) 568-3720"
+  },
+  {
+    "id": "1c344192-b88d-4696-ba0b-d99260867a05",
+    "name": "Reuben Henry",
+    "email": "pharetra.ut@dictum.co.uk",
+    "phone": "(715) 598-5792"
+  },
+  {
+    "id": "1c344192-b88d-4696-ba0b-d99260867a06",
+    "name": "Simon Morton",
+    "email": "dui.Fusce.diam@Donec.com",
+    "phone": "(233) 738-2360"
+  },
+  {
+    "id": "1c344192-b88d-4696-ba0b-d99260867a07",
+    "name": "Thomas Lucas",
+    "email": "nec@Nulla.com",
+    "phone": "(704) 398-7993"
+  },
+  {
+    "name": "Dona Veb",
+    "email": "dona.veb@gmail.com",
+    "phone": "(077) 888-7788",
+    "id": "742588cb-dd92-49f0-9224-30c5c5b910a6"
+  }
+]
