+++ model/contacts/listContacts.js
@@ -1,7 +1,7 @@
 const fs = require('fs/promises')
 
 const listContacts = async() => {
-  const data = await fs.readFile('./model/contacts.json')
+  const data = await fs.readFile('./db/contacts.json')
   const contacts = JSON.parse(data)
   return contacts
 }
