+++ model/contacts/removeContact.js
